pyproject.toml
[project]
name = "unit-cooler"
version = "0.1.0"
description = "エアコンが動作を開始すると自動的に室外機へのミスト噴射を行うシステムです。"
authors = [
    { name = "KIMATA Tetsuya", email = "kimata@green-rabbit.net" }
]

readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "rpi-lgpio>=0.6",
    "flask>=3.1.1",
    "flask-cors>=5.0.1",
    "docopt-ng>=0.9.0",
    "influxdb-client[ciso]>=1.49.0",
    "my-lib @ git+https://github.com/kimata/my-py-lib@f7251a5d834cf8201f59e367f94871800f000c01",
    "spidev>=3.7",
    "fluent-logger>=0.11.1",
    "pyzmq>=27.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "scipy>=1.10.0",
    "pillow>=11.2.1",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "flaky>=3.8.1",
    "playwright>=1.57.0",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-html>=4.1.1",
    "pytest-mock>=3.15.1",
    "pytest-playwright>=0.7.2",
    "pytest-xdist>=3.8.0",
    "pytest-timeout>=2.4.0",
    "time-machine>=3.2.0",
    "mypy>=1.19.1",
    "pyright>=1.1.407",
    "types-pytz>=2025.2.0",
    "types-paramiko>=4.0.0",
    "types-pyyaml>=6.0.12",
]

[tool.uv]
default-groups = ["dev"]

[tool.mypy]
warn_return_any = false
warn_unused_configs = true
ignore_missing_imports = true
python_version = "3.12"
exclude = [
    "^tests/",
    "^.venv/",
]

[tool.pyright]
venvPath = "."
venv = ".venv"
include = ["src", "tests"]
reportMissingImports = false
reportArgumentType = false
reportOperatorIssue = false
reportIndexIssue = false
reportGeneralTypeIssues = false
reportAttributeAccessIssue = false
reportPrivateImportUsage = false
reportOptionalSubscript = false
reportOptionalMemberAccess = false
reportFunctionMemberAccess = false
reportCallIssue = false
reportReturnType = false
reportAssignmentType = false
reportPossiblyUnboundVariable = false

[tool.ruff]
line-length = 110

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP"]
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/unit_cooler"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.pytest.ini_options]
minversion = "6.0"
# E2E テストはデフォルトで除外（外部サーバーが必要なため）
# E2E テスト実行: uv run pytest tests/e2e/ --host <host> --port <port>
addopts = "--verbose --timeout=300 --durations=10 --log-file-level=DEBUG --log-format=\"%(asctime)s %(levelname)s %(message)s\" --log-format=\"%(asctime)s %(levelname)s [%(filename)s:%(lineno)s %(funcName)s] %(message)s\" --capture=sys --junit-xml=reports/junit-report.xml --html=reports/pytest.html --self-contained-html --cov=src --cov-report=html --cov-report=term-missing --cov-report=xml:reports/coverage.xml --cov-report=lcov:reports/coverage.lcov --numprocesses=auto --ignore=tests/e2e"
testpaths = ["tests"]
filterwarnings = [
    "ignore:datetime\\.datetime\\.utcfromtimestamp\\(\\) is deprecated:DeprecationWarning",
    "ignore::DeprecationWarning:multiprocessing\\.popen_fork",
    # pydub の警告（Python 3.12+ でのエスケープシーケンスと audioop 非推奨）
    "ignore:invalid escape sequence:SyntaxWarning:pydub.*",
    "ignore:'audioop' is deprecated:DeprecationWarning:pydub.*",
    "ignore:unclosed database.*:ResourceWarning:rich.*",
    "ignore:unclosed database.*:ResourceWarning:influxdb_client.*",
    "ignore:unclosed database.*:ResourceWarning:coverage.*",
    "ignore:unclosed database.*:ResourceWarning:time_machine.*",
    "ignore:unclosed database.*:ResourceWarning:werkzeug.*",
    "ignore:unclosed database.*:ResourceWarning:_pytest.*",
    "ignore:unclosed database.*:ResourceWarning:typing.*",
    "ignore:unclosed database.*:ResourceWarning:multiprocessing.*",
    "ignore:unclosed database.*:ResourceWarning:logging.*",
    "ignore::ResourceWarning:_pytest.unraisableexception",
    "ignore::ResourceWarning:coverage.sqldata",
    "ignore::ResourceWarning:coverage.collector",
]

[tool.coverage.run]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

[tool.coverage.html]
directory = "reports/coverage"
